Implement `Uffd::read_events`.
This commit implements `Uffd::read_events` that users can call to read multiple
events into a caller-supplied buffer.

An iterator is returned that reads the `Event` objects out of the buffer for
the events that were read.

The implementation of `read_event` was updated as a call to `read_events`,
supplying a buffer large enough for a single event.

Three tests were added: `read_event`, `nonblocking_read_event`, and `read_events`.
The `read_events` test is inherently racy, but is implemented to help minimize
hitting the race condition.