Implement uffd-wp support.
This commit implements the Linux 5.7 `uffd-wp` feature that enables
protecting a range of memory pages as "write protected" and receiving
events in user space about any writes to those pages.

The `linux5_7` feature was added to enable the support.

A test was added to demonstrate how to use the feature.