Fix a file descriptor leak when creation fails
This commit fixes an issue where the uffd file descriptor is created but
might be leaked if creation of `Uffd` fails in subsequent syscalls
afterwards. The fix here is to wrap up the raw integer file descriptor
in a `Uffd` as soon as it's available, allowing the destructor to close
the file descriptor if an early-exit happens.