Fix linux4_14 feature flags
Before this change, code intended for the `linux4_14` feature would not get compiled in as expected. The reason is that the feature flags were not specified in the way Rust expects (this is a common footgun).

This change fixes the `linux4_14` feature flag usage, and safely casts the `ptid` to the expected `nix::unistd::Pid`.