use /dev/userfaultfd when present
Linux kernel 6.1 introduces a /dev/userfaultfd pseudo-device which
allows creating a UFFD objects by issuing USERFAULTFD_IOC_NEW ioctl() to
the device [1]. This way, a process does not need to have the
CAP_SYS_PTRACE capability for creating UFFD objects that can handle
kernel-triggered page faults. UFFDs created through this interface can
always handle kernel-triggered page faults. Access to the device is
granted through normal filesystem permissions to the device file.

This commit changes the way we create Uffd objects, to first try to
create the file descriptor through /dev/userfaultfd. If the file exists
we will try to create the descriptor using ioctl in this device. If the
device does not exist, we will fall back to the syscall.

Signed-off-by: Babis Chalios <bchalios@amazon.es>