test: fix tests for newer kernels
The ubuntu-latest kernel now has a 6.2 kernel. That makes the creation
of a new userfault file descriptor to use path that creates the fd
issuing an ioctl to /dev/userfaultfd. However, in our Github action
/dev/userfaultfd is read/write only by root whereas the job is running
as the "runner" user. This causes the tests to fail, because they can't
open the device file to issue the ioctl.

This commit does two things. First it creates a test matrix so that we
run the github action both on ubuntu-latest (with a newer kerenl) and
ubuntu-20.04 that uses a 5.15 kernel. That ensures that we test both
file descriptor creation paths (/dev/userfaultfd and userfaultfd
syscall). Second, in the case of ubuntu-latest it gives to the "runner"
user access to /dev/userfaultfd file.

Signed-off-by: Babis Chalios <bchalios@amazon.es>